refactor(desktop): organize support and editor hooks
