+++ desktop/src/components/workspace/chat/tool-calls/FileChangeCall.tsx
@@ -11,7 +11,7 @@ import {
   FilePlus,
   Minus,
 } from "@/components/ui/icons";
-import { useOpenInDefaultEditor } from "@/hooks/editor/use-open-in-default-editor";
+import { useOpenInDefaultEditor } from "@/hooks/editor/workflows/use-open-in-default-editor";
 import { useWorkspacePath } from "@/providers/WorkspacePathProvider";
 import { TOOL_CALL_BODY_MAX_HEIGHT_CLASS } from "@/lib/domain/chat/tools/tool-call-layout";
 import { ToolActionDetailsPanel } from "./ToolActionDetailsPanel";
