+++ desktop/src/components/workspace/chat/tool-calls/ToolFileChip.tsx
@@ -4,8 +4,8 @@ import { FileTreeEntryIcon } from "@/components/ui/file-icons";
 import { PopoverButton } from "@/components/ui/PopoverButton";
 import { PopoverMenuItem } from "@/components/ui/PopoverMenuItem";
 import { Copy, ExternalLink } from "@/components/ui/icons";
-import { useFilePathNativeContextMenu } from "@/hooks/editor/use-file-path-native-context-menu";
-import { useOpenInDefaultEditor } from "@/hooks/editor/use-open-in-default-editor";
+import { useFilePathNativeContextMenu } from "@/hooks/editor/ui/use-file-path-native-context-menu";
+import { useOpenInDefaultEditor } from "@/hooks/editor/workflows/use-open-in-default-editor";
 import { useWorkspacePath } from "@/providers/WorkspacePathProvider";
 
 interface ToolFileChipProps {
