+++ desktop/src/components/workspace/shell/sidebar/MainSidebar.tsx
@@ -39,7 +39,7 @@ import { useCloudAvailabilityState } from "@/hooks/cloud/use-cloud-availability-
 import { useCloudBilling } from "@/hooks/cloud/use-cloud-billing";
 import { useCloudRepoConfigs } from "@/hooks/access/cloud/use-cloud-repo-configs";
 import { useDebugRenderCount } from "@/hooks/ui/use-debug-render-count";
-import { useSidebarSupportContext } from "@/hooks/support/use-sidebar-support-context";
+import { useSidebarSupportContext } from "@/hooks/support/derived/use-sidebar-support-context";
 import { useSessionSelectionStore } from "@/stores/sessions/session-selection-store";
 import { useWorkspaceUiStore } from "@/stores/preferences/workspace-ui-store";
 import { useWorkspaceDisplayNameActions } from "@/hooks/workspaces/use-workspace-display-name-actions";
