+++ desktop/src/hooks/editor/ui/use-file-path-native-context-menu.ts
@@ -10,6 +10,7 @@ export function useFilePathNativeContextMenu({
   onOpen: () => void;
   onCopy: () => void;
 }) {
+  // Product-specific native context menu wiring for file-path UI surfaces.
   return useNativeContextMenu(() =>
     buildFilePathNativeContextMenuItems({
       canOpen,
