+++ desktop/src/hooks/editor/workflows/use-open-in-default-editor.ts
