+++ desktop/src/hooks/support/derived/use-sidebar-support-context.ts
